Generate specific gambit config for CI
