+++ package.json
@@ -21,6 +21,7 @@
     "test:e2e:local-env": "yarn hardhat test test-e2e/*",
     "test:storage": "./scripts/storage_layout_test.bash",
     "test:signatures": "./scripts/signatures_test.bash",
+    "prepare:mutation:ci-config": "ts-node test-mutation/generateGambitConfig.ts",
     "test:mutation": "ts-node test-mutation/gambitTester.ts",
     "test:unused:errors": "./test/unused-errors/find_unused_errors.sh",
     "deploy:local:token-bridge": "ts-node ./scripts/local-deployment/deployCreatorAndCreateTokenBridge.ts",
